Show percentage of total time on profiled method

+++ haju3d-common/src/main/java/fi/haju/haju3d/util/ProfileInterceptor.java
@@ -24,7 +24,7 @@ public class ProfileInterceptor implements MethodInterceptor {
     @Override
     public String toString() {
       long totTime = System.currentTimeMillis() - startTime;
-      return method + "\t" + (time / calls) + "\t" + calls + "\t" + time + "\t" + (time / totTime);
+      return method + "\t" + (time / calls) + "\t" + calls + "\t" + time + "\t" + (time / (double)totTime);
     }
   }
